+++ packages/js/e2e-core-tests/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Unreleased
 
+## Changed
+
+- Updated top level menu css selectors
+
+## Fixed
+- Moved `merchant.login()` out of `beforeAll()` block and into test body for retried runs.
+
 ## Added
 
 - A `specs/data` folder to store page element data.